linkedin_finder/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
pandas>=2.2.0

